chore(@clayui/drop-down): removes the useDropdownCloseInteractions hook and moves impl to the component
